AB: use double instead of float

+++ source/drawing.h
@@ -5,20 +5,20 @@
 
 unsigned int add_file_location(const char*);
 void calculate_coordinates_on_graph(int);
-int calculate_number_length(float);
+int calculate_number_length(double);
 void cleanup();
 void draw_grid(ImDrawList*);
 size_t estimated_label_row_number();
 void general_transform_coordinates(unsigned int*, unsigned int, unsigned int, int, bool = false);
-void general_transform_coordinates(float*, float, int, int, bool = false);
+void general_transform_coordinates(double*, double, int, int, bool = false);
 void init_graph_window(bool*);
 void init_movie(bool*);
 void init_settings_menubar();
 void init_settings_window(bool*);
 void init_video_window(bool*);
 int init_window();
-void max_stats(unsigned int*, float*);
-void min_stats(unsigned int*, float*);
+void max_stats(unsigned int*, double*);
+void min_stats(unsigned int*, double*);
 void pop_disable();
 void push_disable();
 void read_image(GLuint*, const char*);
@@ -28,8 +28,8 @@ void set_video_buttons_location();
 bool show_at_least_one_stat_data();
 void show_help_marker(const char*);
 void start_main_loop();
-void trensform_movie_coordinates(float *x, float *y);
-void trensform_distance(float *r);
+void trensform_movie_coordinates(double *x, double *y);
+void trensform_distance(double *r);
 void zoom();
 
 #endif /* drawing_h */
